backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
motor==3.1.2
pymongo==4.3.3
pydantic==1.10.7
python-multipart==0.0.6
boto3==1.26.135
qdrant-client==1.1.7
sentence-transformers==2.2.2
PyPDF2==3.0.1
httpx==0.24.0
email-validator==2.0.0
celery==5.2.7
redis==4.5.5
python-dotenv==1.0.0
